feat(as-4808 moved functionality into enter-code): as-4808 moved functionality into enter-code
