+++ packages/forms-web-app/src/controllers/full-appeal/submit-appeal/enter-code.js
@@ -9,10 +9,25 @@ const {
 	}
 } = require('../../../lib/full-appeal/views');
 
+const { calculateDeadline } = require('../lib/calculate-deadline');
+const {
+	VIEW: {
+		FULL_APPEAL: { CANNOT_APPEAL }
+	}
+} = require('../lib/full-appeal/views');
+
 const getEnterCode = async (req, res) => {
 	const {
 		session: { appeal }
 	} = req;
+	const deadlineHasPassed = calculateDeadline.hasDeadlineDatePassed(
+		appeal.decisionDate,
+		appeal.appealType,
+		appeal.eligibility.applicationDecision
+	);
+	if (deadlineHasPassed) {
+		res.redirect(`/${CANNOT_APPEAL}`);
+	}
 	const url = `/${REQUEST_NEW_CODE}`;
 	await sendToken(appeal);
 	res.render(ENTER_CODE, {
